+++ tests/test_authorize_and_void.py
@@ -3,7 +3,6 @@
 from __future__ import absolute_import
 
 import logging
-from decimal import Decimal
 from braspag.consts import PAYMENT_METHODS
 from braspag import BraspagRequest
 from .base import BraspagTestCase
@@ -22,7 +21,7 @@ class AuthorizeAndVoidTest(BraspagTestCase):
                                      'customer_name': u'José da Silva',
                                      'customer_email': 'jose123@dasilva.com.br',
                                      'transactions': [{
-                                         'amount': Decimal(1000),
+                                         'amount': float(1000),
                                          'card_holder': 'Jose da Silva',
                                          'card_number': '0000000000000001',
                                          'card_security_code': '123',
@@ -35,7 +34,7 @@ class AuthorizeAndVoidTest(BraspagTestCase):
 
     def _authorize_callback(self, response):
         assert response.success == True
-        assert response.transactions[0]['amount'] == Decimal('1000.00')
+        assert response.transactions[0]['amount'] == float('1000.00')
         assert response.order_id == u'2cf84e51-c45b-45d9-9f64-554a6e088668'
         assert response.correlation_id == u'782a56e2-2dae-11e2-b3ee-080027d29772'
         assert response.transactions[0]['payment_method'] == 997
@@ -54,7 +53,7 @@ class AuthorizeAndVoidTest(BraspagTestCase):
 
     def _void_callback(self, response):
         assert response.success == True
-        assert response.transactions[0]['amount'] == Decimal('1000.00')
+        assert response.transactions[0]['amount'] == float('1000.00')
         assert response.correlation_id == u'782a56e2-2dae-11e2-b3ee-080027d29772'
         assert response.transactions[0]['return_code'] == '0' # TODO: transformar em constante
         assert response.transactions[0]['return_message'] == u'Operation Successful'
