+++ tests/test_get_order_id_by_transaction_id.py
@@ -2,7 +2,6 @@
 
 from __future__ import absolute_import
 
-from decimal import Decimal
 from braspag.consts import PAYMENT_METHODS
 from braspag import BraspagRequest
 from .base import BraspagTestCase
@@ -21,7 +20,7 @@ class GetOrderIdTest(BraspagTestCase):
                                      'customer_name': u'José da Silva',
                                      'customer_email': 'jose123@dasilva.com.br',
                                      'transactions': [{
-                                         'amount': Decimal(1000),
+                                         'amount': float(1000),
                                          'card_holder': 'Jose da Silva',
                                          'card_number': '0000000000000001',
                                          'card_security_code': '123',
@@ -34,7 +33,7 @@ class GetOrderIdTest(BraspagTestCase):
 
     def _authorize_callback(self, response):
         assert response.success == True
-        assert response.transactions[0]['amount'] == Decimal('1000.00')
+        assert response.transactions[0]['amount'] == float('1000.00')
         assert response.order_id == u'2cf84e51-c45b-45d9-9f64-554a6e088668'
         assert response.correlation_id == u'782a56e2-2dae-11e2-b3ee-080027d29772'
         assert response.transactions[0]['payment_method'] == 997
